events: remove packet error chan logic
When an error is received in the event loop, the correct thing to do is
just return. We used to need some different error handling due to the
use of read deadlines, but since that is gone, there is no need to have
this logic.

This also simplifies the logic in eventTransportCommunicate, because we
can just send(), and then wait for the response on el.pc, returning an
error if the channel is closed.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>